Phone number validation in R

dialr is an R interface to Google's libphonenumber library.

Overview

dialr is an R port of Google's libphonenumber library. It uses the java implementation of libphonenumber via rJava for all phone number processing.

For a full rundown of libphonenumber see their github and javadocs.

Options

dialr.home: The default region used to process phone numbers where no region is provided. (default: "AU").

dialr.format: The default format used to print clean phone numbers. See format.phone() for details. (default: "E164").

dialr.show_progress: Should lengthy operations such as phone() show a progress bar? (default: TRUE).

See also

Examples

library(dialr)

# Parse phone number vector
x <- c(0, 0123, "0404 753 123", "61410123817")
x <- phone(x, "AU")

is_parsed(x)    # Was the phone number successfully parsed?

#> [1] FALSE  TRUE  TRUE  TRUE

is_valid(x)     # Is the phone number valid?

#> [1] FALSE FALSE  TRUE  TRUE

is_possible(x)  # Is the phone number possible?

#> [1] FALSE FALSE  TRUE  TRUE

get_region(x)   # What region (ISO country code) is the phone number from?

#> [1] NA   NA   "AU" "AU"

get_type(x)     # Is the phone number a fixed line, mobile etc.

#> [1] NA        "UNKNOWN" "MOBILE"  "MOBILE" 

format(x)

#> [1] NA             "+61123"       "+61404753123" "+61410123817"

format(x, home = "AU")

#> [1] NA           "123"        "0404753123" "0410123817"