Phone number types

Functions for working with phone number types.

get_type(x)

get_supported_types()

get_types_for_region(x)

Arguments

x
A phone vector, or a character vector of ISO country codes

Details

In addition to validity, libphonenumber can identify phone number type - it is able to distinguish Fixed-line, Mobile, Toll-free, Premium Rate, Shared Cost, VoIP, Personal Numbers, UAN, Pager, and Voicemail (whenever feasible).

get_type(x) returns the phone number type for each element of a phone vector.

Valid phone number types differ by region. get_types_for_region(x) returns a list of character vectors of valid types for each provided ISO country code. Use get_supported_types() to see a full list of supported types.

libphonenumber reference

get_type(): PhoneNumberUtil.getNumberType()

get_supported_types(): PhoneNumberUtil.PhoneNumberType

get_types_for_region(): PhoneNumberUtil.getSupportedTypesForRegion()

See also

Examples

  x <- phone(c(0, 0123, "0412 345 678", "61412987654", "03 9123 4567", "+12015550123"), "AU")
  get_type(x)

#> [1] NA                     "UNKNOWN"              "MOBILE"              
#> [4] "MOBILE"               "FIXED_LINE"           "FIXED_LINE_OR_MOBILE"

  get_supported_types()

#>  [1] "FIXED_LINE"           "MOBILE"               "FIXED_LINE_OR_MOBILE"
#>  [4] "TOLL_FREE"            "PREMIUM_RATE"         "SHARED_COST"         
#>  [7] "VOIP"                 "PERSONAL_NUMBER"      "PAGER"               
#> [10] "UAN"                  "VOICEMAIL"            "UNKNOWN"             

  get_types_for_region("AU")

#> [[1]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"         "PAGER"       
#> 

  get_types_for_region(c("GB", "US"))

#> [[1]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "VOIP"            "PERSONAL_NUMBER" "PAGER"           "UAN"            
#> 
#> [[2]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER" "UAN"            
#> 

  get_types_for_region(get_supported_regions())

#> [[1]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[2]]
#> [1] "FIXED_LINE"  "MOBILE"      "TOLL_FREE"   "SHARED_COST"
#> 
#> [[3]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "VOIP"            "PERSONAL_NUMBER" "UAN"            
#> [9] "VOICEMAIL"      
#> 
#> [[4]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[5]]
#> [1] "FIXED_LINE" "MOBILE"     "VOIP"       "UAN"       
#> 
#> [[6]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE"  "PAGER"     
#> 
#> [[7]]
#> [1] "FIXED_LINE" "MOBILE"     "UAN"       
#> 
#> [[8]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[9]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "UAN"         
#> 
#> [[10]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[11]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "VOIP"            "PERSONAL_NUMBER" "PAGER"          
#> 
#> [[12]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[13]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "PERSONAL_NUMBER"
#> 
#> [[14]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"        
#> 
#> [[15]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[16]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "UAN"         
#> 
#> [[17]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[18]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"        
#> 
#> [[19]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> 
#> [[20]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"         "PAGER"       
#> 
#> [[21]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> 
#> [[22]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "UAN"         
#> 
#> [[23]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[24]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "UAN"         
#> 
#> [[25]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "UAN"         
#> 
#> [[26]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "VOIP"            "PERSONAL_NUMBER" "UAN"            
#> 
#> [[27]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "UAN"         
#> 
#> [[28]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE"  "VOIP"      
#> 
#> [[29]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[30]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "UAN"         
#> 
#> [[31]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[32]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[33]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[34]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> 
#> [[35]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[36]]
#> [1] "FIXED_LINE" "MOBILE"     "VOIP"       "UAN"       
#> 
#> [[37]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[38]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[39]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[40]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[41]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "UAN"         
#> 
#> [[42]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[43]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE"  "VOIP"      
#> 
#> [[44]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> 
#> [[45]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE"  "VOIP"      
#> 
#> [[46]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER" "UAN"            
#> 
#> [[47]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[48]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "PERSONAL_NUMBER" "PAGER"           "VOICEMAIL"      
#> 
#> [[49]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[50]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> [6] "UAN"         
#> 
#> [[51]]
#> [1] "FIXED_LINE"   "MOBILE"       "PREMIUM_RATE" "VOIP"        
#> 
#> [[52]]
#> [1] "FIXED_LINE" "MOBILE"     "VOIP"      
#> 
#> [[53]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> 
#> [[54]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "VOIP"            "PERSONAL_NUMBER" "UAN"            
#> [9] "VOICEMAIL"      
#> 
#> [[55]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> 
#> [[56]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"         "PAGER"        "UAN"         
#> 
#> [[57]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[58]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[59]]
#> [1] "FIXED_LINE"   "MOBILE"       "PREMIUM_RATE" "VOIP"        
#> 
#> [[60]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"        
#> 
#> [[61]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[62]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "VOIP"            "PERSONAL_NUMBER"
#> 
#> [[63]]
#> [1] "FIXED_LINE" "MOBILE"     "VOIP"      
#> 
#> [[64]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"        
#> 
#> [[65]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[66]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[67]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[68]]
#> [1] "FIXED_LINE"   "MOBILE"       "PREMIUM_RATE"
#> 
#> [[69]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[70]]
#> [1] "FIXED_LINE"   "MOBILE"       "PREMIUM_RATE"
#> 
#> [[71]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[72]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "PERSONAL_NUMBER" "PAGER"           "UAN"            
#> [9] "VOICEMAIL"      
#> 
#> [[73]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[74]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[75]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> 
#> [[76]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[77]]
#> [1] "FIXED_LINE"  "MOBILE"      "TOLL_FREE"   "SHARED_COST" "VOIP"       
#> 
#> [[78]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[79]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> 
#> [[80]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[81]]
#> [1] "FIXED_LINE"
#> 
#> [[82]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "VOIP"            "PERSONAL_NUMBER"
#> 
#> [[83]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> 
#> [[84]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[85]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[86]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[87]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[88]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> 
#> [[89]]
#> [1] "FIXED_LINE"  "MOBILE"      "SHARED_COST" "PAGER"      
#> 
#> [[90]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"        
#> 
#> [[91]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "PERSONAL_NUMBER" "UAN"            
#> 
#> [[92]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[93]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[94]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "VOIP"            "PERSONAL_NUMBER" "UAN"            
#> [9] "VOICEMAIL"      
#> 
#> [[95]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[96]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[97]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> 
#> [[98]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[99]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER" "PAGER"           "UAN"            
#> 
#> [[100]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER" "VOICEMAIL"      
#> 
#> [[101]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "PERSONAL_NUMBER" "PAGER"           "UAN"            
#> [9] "VOICEMAIL"      
#> 
#> [[102]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[103]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "VOIP"            "PERSONAL_NUMBER" "UAN"            
#> 
#> [[104]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[105]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"        
#> 
#> [[106]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[107]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[108]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[109]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> 
#> [[110]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[111]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"        
#> 
#> [[112]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER" "UAN"            
#> 
#> [[113]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE"  "VOIP"      
#> 
#> [[114]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[115]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[116]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> 
#> [[117]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[118]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[119]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "VOIP"            "PERSONAL_NUMBER"
#> 
#> [[120]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[121]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[122]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "PERSONAL_NUMBER" "UAN"            
#> 
#> [[123]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[124]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[125]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[126]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[127]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> [6] "UAN"         
#> 
#> [[128]]
#> [1] "FIXED_LINE" "MOBILE"     "VOIP"       "UAN"       
#> 
#> [[129]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "UAN"         
#> 
#> [[130]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[131]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[132]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[133]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> 
#> [[134]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"         "UAN"         
#> 
#> [[135]]
#> [1] "FIXED_LINE" "MOBILE"     "VOICEMAIL" 
#> 
#> [[136]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[137]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "VOIP"            "PERSONAL_NUMBER" "PAGER"           "UAN"            
#> 
#> [[138]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[139]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[140]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE"  "VOIP"      
#> 
#> [[141]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[142]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "VOIP"            "PERSONAL_NUMBER" "PAGER"           "UAN"            
#> 
#> [[143]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[144]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[145]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE"  "VOIP"      
#> 
#> [[146]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[147]]
#> [1] "FIXED_LINE" "MOBILE"     "VOIP"      
#> 
#> [[148]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[149]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[150]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "PERSONAL_NUMBER"
#> 
#> [[151]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[152]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[153]]
#> [1] "FIXED_LINE" "MOBILE"     "VOIP"      
#> 
#> [[154]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[155]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[156]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER" "PAGER"           "UAN"            
#> 
#> [[157]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[158]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER" "UAN"            
#> 
#> [[159]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[160]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE"  "VOIP"      
#> 
#> [[161]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> [6] "UAN"         
#> 
#> [[162]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "UAN"         
#> 
#> [[163]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[164]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "VOIP"            "PERSONAL_NUMBER" "UAN"            
#> [9] "VOICEMAIL"      
#> 
#> [[165]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"         "UAN"          "VOICEMAIL"   
#> 
#> [[166]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "VOIP"            "PERSONAL_NUMBER" "UAN"            
#> 
#> [[167]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "UAN"         
#> 
#> [[168]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[169]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"         "UAN"         
#> 
#> [[170]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[171]]
#> [1] "FIXED_LINE" "MOBILE"     "VOIP"       "UAN"       
#> 
#> [[172]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> [6] "UAN"          "VOICEMAIL"   
#> 
#> [[173]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "VOIP"            "PERSONAL_NUMBER"
#> 
#> [[174]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[175]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "VOIP"            "PERSONAL_NUMBER" "PAGER"           "UAN"            
#> 
#> [[176]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE"  "VOIP"      
#> 
#> [[177]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[178]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "PERSONAL_NUMBER" "PAGER"           "UAN"            
#> 
#> [[179]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "VOIP"            "PERSONAL_NUMBER" "PAGER"           "UAN"            
#> 
#> [[180]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[181]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[182]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[183]]
#> [1] "FIXED_LINE" "MOBILE"     "VOIP"      
#> 
#> [[184]]
#> [1] "FIXED_LINE"   "MOBILE"       "PREMIUM_RATE"
#> 
#> [[185]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[186]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[187]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "VOIP"            "PERSONAL_NUMBER" "PAGER"           "UAN"            
#> 
#> [[188]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[189]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER" "PAGER"          
#> 
#> [[190]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "VOIP"            "PERSONAL_NUMBER"
#> 
#> [[191]]
#> [1] "FIXED_LINE" "MOBILE"     "UAN"       
#> 
#> [[192]]
#> [1] "FIXED_LINE"   "MOBILE"       "PREMIUM_RATE" "SHARED_COST" 
#> 
#> [[193]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[194]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "UAN"         
#> [6] "VOICEMAIL"   
#> 
#> [[195]]
#> [1] "FIXED_LINE" "MOBILE"     "UAN"       
#> 
#> [[196]]
#> [1] "FIXED_LINE"   "MOBILE"       "PREMIUM_RATE"
#> 
#> [[197]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[198]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "PERSONAL_NUMBER" "UAN"            
#> 
#> [[199]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"        
#> 
#> [[200]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> 
#> [[201]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[202]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> 
#> [[203]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[204]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"         "UAN"         
#> 
#> [[205]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> [6] "UAN"         
#> 
#> [[206]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[207]]
#> [1] "FIXED_LINE" "MOBILE"     "VOIP"      
#> 
#> [[208]]
#> [1] "FIXED_LINE" "MOBILE"     "VOIP"      
#> 
#> [[209]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> 
#> [[210]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[211]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE"  "VOIP"      
#> 
#> [[212]]
#> [1] "FIXED_LINE" "MOBILE"     "VOIP"      
#> 
#> [[213]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[214]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[215]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[216]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[217]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER"
#> 
#> [[218]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> [6] "PAGER"        "UAN"         
#> 
#> [[219]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> 
#> [[220]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "UAN"         
#> 
#> [[221]]
#> [1] "FIXED_LINE" "MOBILE"     "VOIP"      
#> 
#> [[222]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "PERSONAL_NUMBER"
#> 
#> [[223]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> 
#> [[224]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[225]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> 
#> [[226]]
#> [1] "FIXED_LINE"   "MOBILE"       "PREMIUM_RATE"
#> 
#> [[227]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[228]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[229]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE"  "UAN"       
#> 
#> [[230]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[231]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "VOIP"        
#> [6] "PAGER"        "UAN"         
#> 
#> [[232]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "VOIP"            "PERSONAL_NUMBER" "UAN"            
#> [9] "VOICEMAIL"      
#> 
#> [[233]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[234]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[235]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[236]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER" "PAGER"          
#> 
#> [[237]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[238]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE"
#> 
#> [[239]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "SHARED_COST"     "PERSONAL_NUMBER"
#> 
#> [[240]]
#> [1] "FIXED_LINE" "MOBILE"    
#> 
#> [[241]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE"  "VOIP"      
#> 
#> [[242]]
#> [1] "FIXED_LINE" "MOBILE"     "TOLL_FREE" 
#> 
#> [[243]]
#> [1] "FIXED_LINE"      "MOBILE"          "TOLL_FREE"       "PREMIUM_RATE"   
#> [5] "PERSONAL_NUMBER" "UAN"            
#> 
#> [[244]]
#> [1] "FIXED_LINE"   "MOBILE"       "TOLL_FREE"    "PREMIUM_RATE" "SHARED_COST" 
#> [6] "VOIP"         "PAGER"        "UAN"         
#> 
#> [[245]]
#> [1] "FIXED_LINE" "MOBILE"    
#>