Format a phone number

ph_format is soft-deprecated and will be removed in a future release.

ph_format(phone, country, format = "NATIONAL", home = NULL, clean = FALSE)

Arguments

phone
Character vector of phone numbers to check

country
Character vector of region codes. If length = 1, the same region code is used for all phone numbers

format
Phone number format to use.

home
ISO country code for home country. If provided, numbers will be formatted for dialing from the home country.

clean
Should non-numeric characters be removed? If TRUE, keeps numbers

Value

a character vector of formatted phone numbers.

Examples

x <- c(0, 0123, "0404 753 123", "61410123817", "+12015550123")
ph_format(x, "AU")

#> Warning: ph_format() is soft deprecated as of dialr 0.2.0
#> please use format.phone() instead

#> [1] NA               "123"            "0404 753 123"   "0410 123 817"  
#> [5] "(201) 555-0123"

ph_format(x, "AU", home = "AU") 

#> Warning: ph_format() is soft deprecated as of dialr 0.2.0
#> please use format.phone() instead

#> [1] NA                    "123"                 "0404 753 123"       
#> [4] "0410 123 817"        "0011 1 201-555-0123"