Old one shot phone functions

This set of methods is a relic of the early development of dialr. They are significantly slower than the current approach. They were soft deprecated in 0.2.0, hard deprecated in 0.4.0 and will be entirely removed in the next major release.

All functions have a corresponding function in the current version of dialr. Due to the changes in approach, the new versions of each function are not drop-in replacements and most need a parsed phone vector as input. See the introductory vignette or the phone class documentation for more detail.

ph_valid() -> is_valid()

ph_region() -> get_region()

ph_possible() -> is_possible()

ph_format() -> format.phone()

ph_type() -> get_type()

ph_example() -> get_example()

ph_valid(phone, country, strict = FALSE)

ph_region(phone, country)

ph_possible(phone, country, detailed = FALSE)

ph_format(phone, country, format = "NATIONAL", home = NULL, clean = FALSE)

ph_type(phone, country)

ph_example(country, type = NULL, home = NULL, clean = FALSE)