dialr: Parse, Format, and Validate International Phone Numbers in R

Parse, format, and validate international phone numbers in R using Google's 'libphonenumber' java library, <https://github.com/googlei18n/libphonenumber>.

Options

dialr.home: The default region used to process phone numbers where no region is provided. (default: "AU").

dialr.format: The default format used to print clean phone numbers. See format.phone() for details. (default: "E164").

dialr.show_progress: Should lengthy operations such as phone() show a progress bar? (default: TRUE).

See also