dialr: Parse, Format, and Validate International Phone Numbers

Parse, format, and validate international phone numbers using Google's 'libphonenumber' java library, <https://github.com/google/libphonenumber>.

Options

dialr.home: The default region used to process phone numbers where no region is provided. (default: "AU").

dialr.format: The default format used to print clean phone numbers. See format.phone() for details. (default: "E164").

dialr.locale: The default Java locale used to retrieve localised results in geographical and carrier information (default: "en").

dialr.show_progress: Should lengthy operations such as phone() show a progress bar? (default: TRUE).

See also

Author

Maintainer: Danny Smith danny@gorcha.org

Other contributors:

The Social Research Centre [copyright holder]